Allow larger errors between wind speeds in regular and rotated coordinate systems.
In theory they should be the same but in practice the difference is usually somewhere between 1e-7 and 1e-11. Every once in a while though the error is a bit larger (~0.0015), usually when the wind speed (vector length) is high. Although the absolute difference is higher, the relative difference to the length of the vector stays roughly the same (~0.001% - ~0.0025%). 

Simple testing with somewhat reasonable values for the vector shows that the absolute error does not ever get higher than ~0.0025.