Add function to check missing values.
This function enables sparse matrix like features that we need if and when
Merge() is used in more complex infos. What Merge() does is basically create
a sparse matrix EXCEPT if all input parameters,levels and times are equal.
For example if we Merge() temperature from height/2 and wind speed from heigh/10
we actually get info that contains four elements of which two are missing.

It is yet to be decided whether the missing value information should be contained
within the matrix or should for example iterators know about this and adjust their
size based on this.