Fixing a very subtle bug in grib bitmap unpacking.
Bitmap is an unsigned char array, each bit in a value representing one value in the actual data set. The bitmap size is calculated 
as bitmap length (in bytes) divided by 8 (eight bits in a byte). Then we round up this value since the number of values does
not necessary follow full byte boundaries. 

The bug was in the rounding up: since both the size of the bitmap and size of byte are both ints, the result of divition is 
also and int and therefore always rounded down (ie. decimal digits where discarded). This lead to the situation that at most 
7 last bits of the bitmap were not handled (they were initialized to zero) and the actual data corresponding to these bits was 
discarded (as the bitmap said therefore that the values are missing).