Splitting grid-class in anticipation of future development: Himan most likely needs to support calculation for irregular grids (ie. for a group of points).
This can be implemented quite nicely by detaching the current 'grid' class from actual implementation stuff and transforming it to an interface. Another
class, in this case 'regular_grid', will then implement the functionality that the original grid class had previously. Then later on we can create 'irregular_grid'
class which will implement the same interface but will support irregular grids. The plugin side of himan does not need to know anything about what's happening
behind the scenes.

The biggest downside is that now we have to implement some casting in certain places in order to actually be able to call the relevant functions in the
child classes.