Performance optimizations for hybrid_height. This plugin is notoriously hard to make performant since the calculation is sequential in nature and cannot very easily be parallelized. This commit adds a separate writer thread group to plugin which is used when
a) calculation is truly sequential (ECMWF) AND
b) external packing is used (f.ex. gzip) AND
c) output files are written to either database or single message per file

Writer thread group works so that every time a grid is finished, the main 
thread will add it to cache and the launch a separate thread to finish the
writing to disk. The writer threads are collected in class destructor.

Use case for this is at CloudSigma servers where a 30% increase in performance
is expected.