Add support for limiting the size of himan cache.
Previously the configuration option use_cache could be used to
prevent himan from using cache. This affects the performance pretty
hard since all data fetching needs to go through disk or OS buffers.

A new configuration option cache_limit is created so that we can specify
at maximum how many cache items are kept in memory. Default value is -1,
meaning no limit to cache size. Value 0 is unallowed since if cache needs
to be disabled, use_cache is more efficient option. So valid values for
cache_limit are >= 1. The size of cache in bytes is of course relative
to the size of read grids.

When the cache size exceeds the given limit, the oldest grid from the cache
is remove. Every time the cache returns an element it will update the timer
of that given element. Cache clean routine is only started when new elements
are inserted to cache.