Fix sneaky bug with relative_humidity calculation.
In some occasions himan failed to scale pressure parameter correctly, and it
transformed hPa to MPa, which is obviously incorrect. 

If pressure data was read from cache (operative routine) everything worked 
correctly, since the parameter unit was kUnknownUnit. When source data was 
read from grib (cloudsigma), grib_api library did set the unit to Pa for 
grib2 files (hybrid level > 127) because that's how its defined in the official 
documentation. When himan encountered unit Pa, it scaled it to hPa (and the
bug was triggered).

This fix will remove the unit check from relative humidity code.