Committing a version of cuda-enabled si-plugin. Current version knows how to calculate MaxThetaE with Cuda and it's quite a bit faster that the CPU version, even with relatively small grid sizes (ECMWF europe).
So this is definitely work in progress, and this is just the first commit of many. It's
not nice to split the changes into multiple commits which leave the plugin in a kind
of a mixed state, but already this commit changes too many lines. Maybe git could solve this
issue for us ...