Extended json_parser to read arbitrary parameter options per plugin
json_parser now reads arbitrary parameter configurations and stores the 
configuration in a std::map of string -> vector<pair<string, string>>.

This map is stored in itsPreconfiguredParameters. The key of the map
is the name of the parameter. The pairs in the vector are a combination
of (option name, option value). Option values are stored as strings and
the user has to convert them to the wanted format.

Usage example:

if (itsConfiguration->ParameterExists("wind"))
{
	auto paramOptions = itsConfiguration->GetParameterOptions("wind");
	for (auto && p : paramOptions)
	{
		std::cout << "param name: " << p.first << " value: " << p.second << "\n";
	}
}

User has to check if the parameter exists with ParameterExists - there's
no error checking in GetParameterOptions!

Configuration example:

...
"processqueue" : [
	{
	"leveltype" : "ground",
	"levels" : "0",
	"plugins" : [ { "name" : "test_plugin",
		        "param_list" : 
			[
			  { "name" : "parameterName1", "parameterOption1" : 1.0 },
			  { "name" : "parameterName2", "min" : -1.0, "max" : 2.0 }
			] } ]
	}
]
...

Parameter options are stored in a list named "param_list". Each parameter 
option definition needs to have "name". Not having a name or empty name 
is a checked runtime error. Each parameter option needs to have a nonempty
value, this is also checked.