Adding a new member itsValue2 for level class.
This member is used in those (rare) cases that a level has two values
defined. It is so for example level type 106 (grib1 number). The naming
of the variable is opaque on purpose: we could have the variables named
like "upper" and "lower", but again they might not apply for some other
level context. Also it should be remembered that int the great majority
of cases there is only one value for a level.

To accomodate the changed, a constructor of level class that had a
number variables is removed. The reason for this is that I wanted
to add a constructor for the two-value case, and it will conflict
with the removed constructor (which was never used anyways).