Change cape plugin output for cape and cin parameters.
Cape plugin calculates three cape variants and one cin for
three source data variations, so all together 14 parameters.
Currently these are all different parameters which reserves
a lot of parameter numbers. In this commit the output is
changed so that only four parameters area defined, and the
source data variants are written to different levels:

* source data: surface --> level height/0
* source data: 500m mix --> level heightlayer/0-500
* source data: most unstable --> level maximumthetae/0

The latter two levels are new, and the middle level requires
changes to both database and fmigrib.