Adding support for lambert conformal conical projection.
For this a new class, lambert_conformal_grid has been
introduced. This class has a dependency for gdal headers,
so those will be a himan dependency from now on. To support
writing to querydata, class NFmiGdalArea is added as a
dependency to querydata plugins. The latest version  of
newbase is needed (16.9.20) due to a bug in the earlier
versions.