Fix source parameter names for precipitation_rate and absolute_humidity.
Mixing ratio parameter names from Harmonie have been incorrect in the
database for a long time. The reason for this is that Harmonie writes
two different parameters with same number: in ground level they are
"instant precipitations", in hybrid levels they are mixing ratios.
Finally the support for separating the parameters based on level
has been added to database and grid_to_neons.