Change logging of reading an index file a bit.
Normally when grib is read, grib plugin reports the bandwidth used
(MB/s). When reading from an index, we can't report that because
it's calculated from the size of the file, and the size of the
index != size of the read data. So just report the time it took
to read the index and the actual message. (We could get the size
of the message from grib_api somehow, but maybe that's too much
effort ...)