Fix height search when searched area (level range) is between two hybrid levels (HIMAN-142).
For some modifiers (min/max/minmax/cnt) the height search was limited:
if a vertical height range was totally between two levels, no value
was found. This is of course a problem with small ranges, but also
it produced inaccurate results as the search ignored start and end
boundaries (heights). Other modifiers (mean, find_*) worked correctly.

This commit copies the functionality of mean modifier to other modifiers
as well. By doing that some of the common code has been moved to separate
functions, as the if statements can be rather hairy. All unit tests
pass.