ensemble support for glameps and its rotating members
+ added MaxMissing to ensemble
+ made probability configuration conform to normal json conf naming conventions
+ added max_missing to probability config

Probability calculations are now based on the number forecasts found for
the current step. Normal behavior doesn't change w.r.t for example
ECMWF EPS. For GLAMEPS, if the step has only 22 forecasts, it doesn't
make sense to use 52 as the ensemble size for calculations.

New configuration naming has to be changed for the production configurations.