Two new modifiers: findheight_gt and findheight_lt.
These two are very similar to findheight, which return the
height where a certain value is found (for example temperature).
The new modifiers just return the height where it's less than
or greater than the given value.

This is an important distinction: some times we might be searching
for f. ex. height where total cloud cover > 80%. If the cloudiness
is 100% starting from ground up, the original findheight modifier
would not find the first value!