Add support for lambert vector rotation for windvector plugin.
Wind vector components must be rotated to earth normal (pointing
earth north & east instead of grid equivalents) before calculating
wind direction. The code is now greatly simplified, but still too
convoluted (something must be done in the near future). The problem
of this specific piece of code is that it needs to be executed
both at CPU and GPU.

Also remove vector calculation.