Add support for per-processqueue time definitions (1), and multiple origintimes for single configuration (2).
Change (1):

Per-processqueue time definitions can be defined as such:

"processqueue" : [
  {
    "origintime" : "latest",
    "hours" : "1,2,3",
    "leveltype" : "height",
    ...
  }
]

Note that both origintime and one of the following three must be defined:
* hours
* start_hour, stop_hour, step
* start_minute, stop_minute, step

Change (2):

A new key 'origintimes' has been introduced, which can be used to define
multiple origintimes. Times must be in format yyyy-mm-dd hh24:mi:ss, and
separated with a comma. 'origintimes' can be used in top level or in
processqueue.

Example:

"origintimes" : "1980-01-01 00:00:00,1981-01-01 00:00:00",
"hours" : "3,6",

The dates that are created from this are simply the cartesian product of the
two, ie. the configuration above creates four different valid times.