Reworked vector component rotation to allow interpolation of components.
When interpolating projected vector component data (to other projections or
latlon), the components have to be rotated. This is because the projected
area north most likely doesn't correspond to geographical north. In our
case this applies to stereographic, lambert and rotated latlon projections.

Currently windvector-plugin has the capacity to do this rotations, as it
is needed to calculate wind direction (magnitude is the same whether the
components are in grid-normal or earth-normal format). A new requirement
is to be able to interpolate the vector components themself from say rotated
latlon to regular latlon.

In order to fulfill this requirement, the rotational code has been merged
with interpolation code and moved to himan-lib.

* information of UV grid relativity is added to grid-class
* interpolation code recognizes the most common component parameters
  and automatically tries to rotate if that is required (when changing
  projection)
* the rotation can be requested explicitly from fetched plugin; this is
  what windvector plugin is doing since it needs the components rotated
  but does not change area or projection
* cuda interpolation code moved from himan-plugins to himan-lib (a change
  that was long due)