Add findheight_gt/lt modifier support for luatool and hitool.
These modifiers are exactly the same as findheight, except that
they will also consider that the first value encountered is might
be above / below a threshold.

A comment copied from hitool.h:

---------------------------
erticalHeightGreaterThan() is similar to VerticalHeight(), but when searching
for a value it also considers the situation where a search is started and the value
is encountered in the very first height.

For example when searching for a height where cloudiness is > 50%, regular VerticalHeight()
does not understand the situation where the first value read is already above the threshold
(stratus cloud).
---------------------------