Add code to speed up auxiliary file read.
SUMMARY

Himan auxiliary file read has been changed so that instead of
scanning through all files in search for the correct grid over
and over again, all grids are read into memory once when himan
starts. Doing this will speed up those cases where many auxiliary
files are used:
* A development case with lots source data and single-point
  interpolation went from 144 seconds to 22 seconds
* A full himan-test run went from 11 minutes to 6 minutes

BACKGROUND

Himan data search order is (by default):
1) cache
2) auxiliary files
3) database

Cache and database are very fast to search, because the files
are found with a "hashing" process. Auxiliary files on the other
hand are very slow, because a linear search is used. When a lot
of data is given as input, the linear search will take a long time.
This is a problem with automated tests which already sometimes take
almost 30 minutes to run, but it is also a nuisance with with
development and debugging where you have to mentally context switch
while waiting for a run to finish.

The current grib plugin contains code that supposedly adds all read
auxiliary file data to cache (ie. trying to warm the cache), IFF it
doesn't need to be interpolated. This is a limitation, of course.

IMPLEMENTATION

The new code has the auxiliary file 'slurp' moved to fetcher. This
allows us to effectively read all auxiliary file data from files
at one time, do the necessary interpolation and rotation and store
all files to cache.

Some highlites of the code:
* By default the new behaviour is to read all auxiliary files to cache
  the first time data is fetched.
* If cache has been disabled, himan will revert to earlier behavior
* If land sea mask is applied, new behaviour cannot be used, because
  himan cache is immutable. Himan will detected if this is the case
  and will abort.
* A new configuration class member has been added that determines if
  the behavior should be used (default=yes)
* A new command line option --no-auxiliary-file-full-cache-read can
  be used to revert back to old behaviour
* The rotation and interpolation is done using 8 parallel threads.
  Threading is done using async/futures.
* The 'slurping' of auxiliary files is done only once, this is
  enforced with a std::once_flag
* The problem of changing grib producer id's arises. ECMWF updates
  generatingProcessIdentifier every time model is updated, we only
  have the current value in database. The new code recognizes a few
  old numbers, but hard coding this information might not be the best
  way to do it.
* Code will unpack any packed data because the current cache does
  not allow packed data.
* CSV plugin has been changed to assume that data given as auxiliary
  file is from the correct producer. I think this is a reasonable
  assumption, given that we have no way of verifying it (csv doesn't
  contain producer ids).