Temporary fix for raw_time Adjust function.
boost::date_time has snap to end of month functionality that
does not work with the way himan is supposed to work. For
example subtracting one year from 2017-02-28 results in
2016-02-29.

This commit will make a temporary fix for the leap year case,
and will remove the support for month resolution in Adjust().

A more permanent fix would be to use some other datetime library,
since the boost behavior is only controllable with a compile time
flag.