Optimization for CAPE MoistLift.
CAPE plugin does a lot of MoistLift, ie. integrates a saturated
air parcel upwards in the atmosphere. This is very compute intensive
especially without GPU's. In CloudSigma the problem is exaggerated due
to the large grid size compared to Kumpula processing.

This commit will add simple threading for MoistLift calculation, where
the whole grid is split into four parts and each part is given to a
separate thread. This is done only for the CPU code. Initial tests
confirm that after change LFC calculation is roughly twice and CAPE
calculation third faster than previously. The MoistLift function in
itself is not modified at all.

The number of 'worker' threads (4) is derived using Stetson-Harrison
method based primarily on the CloudSigma VM CPU count.