Fix situation where LFC level was found but EL level was not
This situation was encountered when LFC was found on the same spot
where starting values where found, and when raising parcel it cooled
faster than the environment. In other words, the env temperature and
parcel temperature curves shared a start point, but diverged from there
and did not converge higher up.

In some other programs the issue is partly solved so that CAPE integration
is started below LFC, which means that there is more vertical ground
to search for CAPE values. This might be worthwhile implementing to
Himan as well. As a workaround, this commit checks for the situation
where LFC is found very close to a hybrid level value, because in those
situations moist lift curve is flat and equals to linear interpolation.
If this case is found, a very small temperature value is added to parcel
so that a miniscule CAPE area is formed.