Add function MoistLiftA_(), an approximation of moist lift
This function uses the "Wobus function" to approximate the
lifted parcel value. This is a polynomial approximation of the
moist adiabatic lapse rate developed by Herman Wobus.

More information can be found with google (though no "official"
docs seems to exist), for example:

http://www.caps.ou.edu/ARPS/arpsbrowser/arps5.2.4browser/html_code/adas/mthermo.f90.html

In the same commit some bugs found in the Tw() function are squashed,
and now Tw() it is more accurate than the Wobus function in warm
conditions. In cold conditions the initial approximation is still
some tenths of a degree off, which is probably caused by the
implementation. I'll add some comparison to cape-plugin documentation,
as this approximation is very useful there.