Fix issue where cape GPU results differed from CPU results.
In cape-plugin, some of the code is shared between CPU and GPU code paths
and some code is platform-depended. CPU code paths got some bugfixes which
were not ported to GPU code. This was not noticed in regression tests
because the test called exit(0) even if results differed.

Code has now been synchronized to both architectures. Also the situation
where a parcel enters CAPE zone is now considered with improved accuracy:
previously it was assumed the the parcel entered the zone halfway between
previous height and current height; now the exact location is used.