Add logging when sticky param cache fails to deliver.
When stcky param cache has information about some producer
and that data is no found when read (for example file has been
already removed), the default behavior is to read through all
producers just to be sure.

A new log message has been added to indicate this situation:
otherwise it's a bit strange to see Himan try to read the same
file twice.