Fix issue where fetcher did and extra cache turn-around for no reason.
When fetcher was able to retrieve data from cache, in a later stage
it tried to push that data back to cache as it did not now at that
stage anymore where the data came from.

That is now changed so that the source is recorded and interpolation,
rotation and cache-push is not done if the data was fetched from cache.