Add support for per-station probability limits
Limits can be set either in the configuration file, or in the database.

Limits are set in the configuration file the following way:
-----------------------------------------------------------

Single threshold, grid based data:

"threshold" : "0.0"

So no change to current way of definining a threshold. This way will
only work with gridded data!

A probability limit for a station is defined like this:

"threshold_123456" : "0.0"

Here 123456 is the station id found in table stations (radon).

A more complete example:

{
	"name" : "probability",
	"param_list" :  [
		{
			"name" : "PROB-WATLEV-LOW-1",
			"input_param1" : "WATLEV-CM",
			"threshold_134254" : "-70",
			"threshold_134248" : "-80"
		},
		{
			"name" : "PROB-WATLEV-HIGH-1",
			"input_param1" : "WATLEV-CM",
			"threshold_134254" : "115",
			"threshold_134248" : "110"
		}
	]
}

Limits are set in the database the following way:
-------------------------------------------------

                 Table "public.station_probability_limit"
      Column       |           Type           |           Modifiers
-------------------+--------------------------+----------------------------------
 id                | integer                  | not null default nextval('...')
 station_id        | integer                  | not null
 param_id          | integer                  | not null
 probability_limit | numeric                  | not null
 last_updater      | text                     |
 last_updated      | timestamp with time zone |

radon=> insert into station_probability_limit values (default, 134254, 1053, 170);
INSERT 0 1

The values in the configuration file will override database. Threshold for gridded
data cannot be set in the database.

NB! A note concerning CSV file structure:

Himan supports currently the following csv structure:

// 0 producer_id
// 1 origin time
// 2 station_id
// 3 station_name
// 4 longitude
// 5 latitude
// 6 param_name
// 7 level_name
// 8 level_value
// 9 level_value2
// 10 forecast period
// 11 forecast_type_id
// 12 forecast_type_value
// 13 value

On the other hand, previ_to_radon.py supports a slightly different structure:
param_name and level_name are replaced with param_id and level_id. The reasoning
behind this is that Himan can work without a database, so it must know all the
values in the csv without accessing database, whereas previ_to_radon.py does --
by definition -- work with radon database. previ_to_radon.py also uses id's
because the database table uses id's and therefore if the data that's inserted
to database is in the exact same format as the database table we can use a
high speed upload command (COPY).

Also added support for parameters PROB-WATLEV-LOW-1 and PROB-WATLEV-HIGH-[1-3]