Do proper CIN integration as per the definition of CIN.
Previously Himan stopped CIN integration as soon as first
CAPE area was found -- even if LFC was not reached yet. This
commit will change the integration so that multiple CIN areas
below LFC are taken into account. Also CIN area start and end
heights are interpolated for improved accuracy.

A future development could be to use modifier_plusminusarea for
the integration as that would utilize trapezoidal rule.