New plugin: blend
This commit implements a new plugin: blend. The plugin takes as input
different producer's forecasts defined in the JSON-configuration. A
single parameter is blended per process queue invocation, as defined in
the JSON. The plugin constructs an ensemble from all the forecasts and
a 'blended' forecast as the control forecast of the ensemble.

The blend is calculated as a linear combination of the forecasts and
weights (supplied in a separate plain-text file). An ad-hoc weight
rebalancing is employed in the case of missing forecast values: Weight
of a missing forecast is distributed equally to the remaining values.