'precision' is now a property of param
Precision describes how many decimal places should be used
to define the accuracy of the parameter when writing to file.
For example value 0 means that all data is rounded to nearest
natural number.

This information is stored in radon database table param_precision.
Previously only grib-plugin accessed the information, now the
property has been promoted to became a part of the param class.

The precision value is *not* used in param <--> param comparisons!