Implement lazy loading of plugins on startup
Himan consists of around 50 plugins implemented as shared
libraries. On Himan startup every one of these is read from
disk, even if it is not used. In a longer Himan execution it
doesn't matter, because the time spent is negligent compared
to overall time, but in small single-plugin executions it is
not necessary to read all plugins.

This commit implements a lazy loading: each plugin is loaded
only when requested.