Add asynchronous execution mode for plugins.
By default Himan plugins are executed in a serialized fashion,
mostly because many of the plugins are dependent on the output
of others. Some plugins, however, lie necessary at the end of the
execution line (e. in a b-tree they are the leaf), and when those plugins
are placed in the critical chain of execution they might unnecessarily
slow down the the total execution of Himan.

A new configuration file key, 'async', has been added. By default
the value is false, confirming to the current behaviour. It can
be set to true for any individual plugin or processqueue element
scope.

Asynchronous execution should only be enabled for those plugins that
have no other plugins as dependants!