Remove compiled_plugin property itsClearTextFormula
This variable was supposed to hold a brief description of what the
plugin does. In reality most of the plugins are using algorithms that
cannot be described so. And in any case code may not be the best
place to document this kind of information -- at least in a variable!