Optimizing hybrid_height plugin by adding prefetching
Hybrid_height for ECMWF data is sequential by nature. We can't really
optimize the algorithm in itself, but we can make sure that data
fetching is not the limiting factor.

This commit adds prefetching of temperature and pressure data that is
done in two separate asynchronous loops. Execution time for ensemble
(51 members, one timestep) is down to from 95 to 45 seconds in dev
environment.