Bugfix: work distribution for hybrid_height plugin
hybrid_height plugin may need to process all levels by one thread.
In the code this is marked as primary dimension being kTimeDimension.
Current way of distribution did not take into account the forecast_type
dimension and therefore processing was done with less parallel threads
than what was possible.

This way of marking the primary dimension is a bit obsolete after
forecast_type was added as a further dimension. I don't have a clear
picture on how to improve it yet; but it is definetly a future todo.