Generalize sparse info functionality.
What this means is that (finally) we can support different level / parameter
combinations naturally without allocating and writing empty grids if the
desired combination of these two is not a cartesian product. This functionality
is required for the upcoming changes in cape and stability plugins.

The latter requires probably some extra explanation. For example
consider the case where we want to calculate temperature for height/2
and dewpoint for height/0 (a made up example).

If we would not have sparse info support, Himan would create the following
grids:

    +------+-------+-------+
    |      |  H/0  |  H/2  |
    +------+-------+-------+
    | T-K  |   X   |   X   |
    +------+-------+-------+
    | TD-K |   X   |   X   |
    +------+-------+-------+

That's four grids allocated and written as files, even if we wanted to
calculate just T-K/H/2 and TD-K/H/0.

With the sparse info support, we get what we expect to get:

    +------+-------+-------+
    |      |  H/0  |  H/2  |
    +------+-------+-------+
    | T-K  |       |   X   |
    +------+-------+-------+
    | TD-K |   X   |       |
    +------+-------+-------+

What "generalization" means is that Himan has had the possibility of doing
this since the beginning, but until now we have not had the use-case for
sparse infos as all data has been just a product of the params and levels.
So this support was never coded, and the de-facto way of doing this was
just the sparse use-case where sparsity = 0.

The default way of creating target parameter info is not changed from the
callers point of view:

SetParams(vector<param>)

is still supported. This function creates a dense info which is a product of
the params in vector and the levels from json file.

In addition to this a new overload has been added:

SetParams(initializer_list<param>, initializer_list<level>)

Similar functions with vector (instead of initializer list) argumens also
exist.

This function creates a sparse info which is a product of the params and
levels given as arguments.

Note that json_parser creates levels for the target info based on the
configuration; if those levels need to be discarded then one should call

info.LevelIterator().Clear()

Sparse info support does require some extra checking every time when when a
grid is accessed because it can be a null pointer (ie an empty entry in the
sparse info).