HIMAN-184: Add support to derive probabilities from a normal distribution
Some parameters like temperature are following the normal (gaussian) distribution.
When the sample size is low, it is beneficial to use the distribution for
determining the probability of reaching some threshold value.

This "gaussian spread" is only valid when a) parameter follows a normal
distribution (currently a hard coded list), and b) the comparison operator is
GTEQ or LTEQ.