Simplifying (*) and extending probability plugin
* Replacing two almost identical functions (Calculate{Normal,Negative}) with a single function
* Remove wind specific function CalculateWind: one should pre-calculate wind speed with windvector
* Add more comparison operators, the available operators are now:

 >= : greater than or equal
 <= : less than or equal
 =  : equal to
 != : not equal to
 <> : not equal to
 =[]: in a set. This can be used to for example check if precipitation form value is within
      a given set. The possible values are given with 'threshold' as a comma-separated list
 [] : inside a range, endpoints included. This can used to for example calculate the probability
      of ideal weather conditions for wind power generation (f.ex. 5 <= FF <= 20). As with
      previous operator, the bounds are given with key 'threshold'.

When doing 'exact comparisons', i.e. operators =, !=, =[] one should remember that the underlying
datatypes are floating point numbers.

The basic comparison operators are picked from <functional> and passed as arguments to Probability()
function. Two custom functors are also implemented, just to show how easy it is to extend the
functionality.

Due to Himan architectural reasons, the implementations have split to several new header files.
Two new files are:
* probability_core.h
* probability_impl.h

File probability_core.h includes the definitions of two structs that are used to pass information from
configuration file to the actual calculating function. The first struct, partial_param_configuration is
parsed from configuration file and multiple instances of the struct are stored to the class member variable
itsParamConfigurations. The struct is called partial because the threshold values are strings and not the
real datatype; the reason for this is that the calculating function is a template, and we can't store a
template variable in a class without making the whole class a template. Struct param_configuration is
'instantiated' from partial_param_configuration when the calculating function is called.

File probability_impl.h contains the required functions for probability calculation. It also contains
the implementations of the two custom functors for =[] and [].

The new implementation supports precipitation form probability calculations, which means we don't need
script preform-prob.lua anymore.

Still to do:
* User input is not validated for multi-value thresholds (=[] or [])
* Station-wise processing does not work with multi-value thresholds, as the database table only supports
  a single value (they could be passed with configuration file); probably the code should reject if
  this type of calculation is defined in configuration file
* Test for operator []

(*) Subjective opinion