STU-6494: Optimize grib packing performance
While examining how to implement grib packing to Himan succesfully I
noticed that each variable precision grib is packed twice: first time
when data is set to grib and second time when we request a precision
change. This was not so evident with smaller gribs but with large
gribs the double-encoding lag is visible.

Another unfortunate thing I noticed is that grib_api seems to have
a bug (or at least some really unexpected behavior) when changing
precision: binary scale factor is always forced to zero after data
repacking. This causes some statistically insignificant but numerically
recognizable jitter to the data. The issue is easily repeated with a
simple test case. What this means that after applying the change each
and every one of our regression test cases will fail :-(