smartsymbol.lua: Calculating three weather-related parameters
WeatherNumber:
An 8-digit number that describes the current weather. All digits are independent
of each other and describe a single parameter.

Example:

12345678
xabcdefg

x (first digit): version number, currently 1
a: thunder probability (number of possible values: 3)
b: potential precipitation form (7)
c: potential precipitation type (3)
d: one hour precipitation sum (8)
e: fog (4)
f: total cloud cover (9)
g: cloud type (8)

The theoretical number of possible permutations of the parameter is:
3 * 7 * 3 * 8 * 4 * 9 * 8 = 145152

SmartSymbol:
For a client application a weather parameter with 145152 possible values is just
unusable. Therefore we 'dumb down' the WeatherNumber to SmartSymbol. This procedure
takes all permutations of one weather state (for example 'Clear skies') and assigns just
one number for it. This reduces the number of different values to 46.

We still leave WeatherNumber as it is, because it can be used when for example
doing interpolation for the data.

Hessaa:
For backwards compatibility, we map WeatherNumber to Hessaa (also called WeatherSymbol3)