Change argument of nan(char*) function used to set Himan missing value
The Cuda device version of nan(char*) has a bug and it does not respect
the given argument but produces always the same nan (0x7fffffff).
Use therefore that value for our missing value nan also; the bug should
be fixed (according to Nvidia, "in a later Cuda release").

More details in HIMAN-189.