Do not re-allocate info memory when pushing to cache
Performance optimization: when info is pushed to cache, it is recreated
so that the memory is also copied. This makes no sense as it doesn't
prevent *later* callers from accessing and modifying the cache data.
Currently we only have a gentlemans agreement that nobody should change
the data in the cache, because it's shared among everybody. So the
optimization is to remove the initial copying of the data. Perhaps a
later TODO item is to somehow code-wise make the cache element const
so it cannot be modified.