Splitting stability plugin into two distinct plugins: 'stability_simple' and 'stability'
This was a semi-hard decision since more plugins means more compilation,
but the current stability plugin produces K-Index which is required for
the older weather indices. Those are calculated *before* calculating
hybrid level height, which is again required by the new parameters in
'stability'.

So therefore the stability calculation that does not required hybrid level
information is now 'stability_simple', and everything else is at 'stability'.

Perhaps later stability_simple can be replaced with a lua-script.