HIMAN-201: Changing cape plugin fully to single precision
All calculations besides 500m mix source data determination on CPU is now
done in single precision. With small areas (like MEPS, Hirlam) the performance
increase is not that great -- around 5 to 15%. The real benefit comes when
dealing with larger areas as global ECMWF grid. The test server does not have
enough memory to run with that area without the server starting to swap.

As single precision support is limited only to this plugin, there's a lot
of conversion from sp to dp and back, which is a bit ugly. Also the himan
cache is still workin in double precision so while the memory usage at GPU
is cut to half, the RAM usage is actually increased (as both dp and sp data
needs to be in memory while the processing is being done).

In order to run lambdas in device code, a couple of new nvcc switches were required:

--expt-extended-lambda and --expt-relaxed-constexpr

Although they are experimental ("--expt"), no NVidia docs I found explained what was
it about them that made it experimental. I found nothing strange while running the
test suites so I'm committing them. If we have problems afterwards we can revert
this decision and implement the same functionality in some other way.